kadm5_randkey_principal interop with Solaris KDC
When kadm5_randkey_principal is called on Solaris KDC (as opposed to
kadm5_randkey_principal_3), the KDC assumes the peer is a Solaris 9
system and only creates keys with DES enctype.

For better interoperability, with this fix the client always tries
calling kadm5_randkey_principal_3 first and only if this procedure is
not present on the remote server, it falls back to calling
kadm5_randkey_principal.