Continue on unknown address error in sendto_kdc
glibc does not declare EAI_NODATA unless _GNU_SOURCE is defined, but
nevertheless returns EAI_NODATA in some circumstances (including some
where it shouldn't, such as when it gets an NXDOMAIN but AI_ADDRCONFIG
was requested).  If we see an unrecognized getaddrinfo result when
resolving a KDC hostname in sendto_kdc, treat it as non-fatal as we
would for EAI_NONAME, and continue on to other KDCs.

ticket: 7526
target_version: 1.12.2
tags: pullup