krb5-config.in: Make --libs spit out LIB_PATH only
krb5-config --libs spits out how MIT Kerberos was linked
explicitly with -rpath, etc. This is not interesting for
the lib user, moreover, it breaks dependends compilations
because it forces every caller to transform his -L to -rpath.
pkg-config files do not do this. They are correct and spit out
-L only. Both should resemble and do resemble now.