Re-encrypt preserved key data in new master key
When we are preserving old key data in kdb_cpw.c, ensure that it is
encrypted with the same master key as the new key data.  This ensures
that the KRB5_TL_MKVNO tl-data on the principal entry applies to all
of the key data, not just some of it.

(cherry picked from commit 32c9b8f1aa1b348388ed227394cc609e68ed833b)

ticket: 7995
version_fixed: 1.13
status: resolved