In ksu, handle typeless default_ccache_name values
When a configured or compiled-in default ccache name doesn't contain a
cache type and ':' as a prefix, add one to the writeable value that we
construct when we go to look it up.  This lets the rest of the
application always assume that it'll be there.

[ghudson@mit.edu: minor style changes]

(cherry picked from commit bda574576a5a2d0613fecf12d820e0adcbedf95c)

ticket: 8007
version_fixed: 1.13
status: resolved