Get rid of static KDC authdata systems
Remove the static_authdata_systems table from kdc_authdata.c and
instead call the relevant functions explicitly in handle_authdata.
Eliminate the flags field from krb5_authdata_systems as it is no
longer used.  Rename the functions to be more descriptive.  Move
simple conditionals on authdata processing to handle_authdata for
clarity.  Move handle_authdata to the end of the file to avoid the
need for static function declarations.