Fix LDAP tests when sasl.h not found
Do not try to run the SASL EXTERNAL auth test if we could not define a
useful interact function.  With current libraries the interact
function is asked for an authorization name, and the bind fails if it
gets an unsuccessful result or if no interaction function is defined.

(cherry picked from commit 8466003864b294cdb9e5547c2f8e574d2c156b13)

ticket: 8049
version_fixed: 1.13.1
status: resolved