Add framework for tracking skipped tests
In k5test.py, add functions skipped() and skip_rest() which output a
message about skipping tests (even without the verbose flag) and also
add a note to the "skiptests" file at the top of the build tree.  In
the top-level make check, empty out skiptests at the beginning and
display it at the end.  Add a subsitution for the skiptests file to
pre.in so that other makefiles can append to it.