Add support for KDC_ERR_MORE_PREAUTH_DATA_REQUIRED (RFC 6113)
This entails some simple changes to the kdc and libkrb5.

On the KDC side, we define the appropriate error code. Additionally,
we allow kdcpreauth plugins to return this error code to clients.

On the libkrb5 side, we treat this error code in the same way as
KDC_ERR_PREAUTH_REQUIRED except that we additionally reset the counter
of tried preauth types, permitting preauth types to continue the converation.