Implement secure cookies
RFC 6113 specifies the ability for the KDC to send a PA-FX-COOKIE
to the client which will be returned on the next AS-REQ. This permits
the KDC to retain state for the subsequent request. In the majority of
cases, this cookie must be encrypted.

This patch enables the transparent encryption/decryption of such
cookies. After the request is decoded, any cookies found are decrypted.
Any cookies set, either internally or by kdcpreauth plugins, will be
automatically encrypted before the reply is encoded.

Additionally, a timestamp is inserted into the cookie before encryption.
This timestamp is validated for freshness during decryption.